pyproject.toml
[project]
name = "rag-knowledge-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "beautifulsoup4>=4.14.3",
    "faiss-cpu>=1.13.1",
    "gradio>=6.1.0",
    "langfuse>=3.11.0",
    "numpy>=2.3.5",
    "openai>=2.13.0",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
]
